Simplify constraints
Summary:
This diff introduces a constraint simplifier. Simplification applies to constraints for a given callable (method or function).

It eliminates all constraints related to an entity if there's a dynamic access. If the entity only has static accesses, then a summary of shape keys along with types are produced. This is only correct while all our analysis is local. For a global analysis, we need to remember if an entity has a dynamic access in function summary. This would be used to invalidate candidate shapes in other functions.

The types for the keys are the union of all types observed for a particular key. This is intended to account for refinements.

Reviewed By: enetsee

Differential Revision: D31576085

fbshipit-source-id: d9f3902e45ad8596d923da37ed4914be77484a41