(3/?) Implement facts_of_decls logic: class attributes
Summary:
This diff implements a straight-forward port of user-attribute decl into facts.

Note:
An attribute of the form `<<Foo('a', 'b'), Bar>>` is turned into
```
user_attributes: [
    UserAttribute {
        name: (
            ...
            "\\Bar",
        ),
        classname_params: [],
    },
    UserAttribute {
        name: (
            ...
            "\\Foo",
        ),
        classname_params: [],
    },
]
```
in decls, whereas we expect to have facts of the form
```
"attributes": {
    "Bar": [],
    "Foo": ["a", "b"],
}
```

In a later diff we'll modify the direct decl parser to preserve the attribute params that facts expects. (See T101762617)

Reviewed By: shayne-fletcher

Differential Revision: D31780923

fbshipit-source-id: c1831b20a12f14031b8493cce956142188dd5092