tighten up datatypes of load failures
Summary:
This diff tightens up the datatypes we use for reporting saved-state load failures:
* use Exception.t, instead of exn+stack
* use Telemetry.t for unstructured data that's destined solely for telemetry

The main effect is that formerly `ServerInit.lazy_saved_state_init` used to exit upon failure and write information into its server_finale_file, but now it does so with richer structured telemetry in that server_finale_file. It will be consumed by the following diff.

Reviewed By: Wilfred

Differential Revision: D31798094

fbshipit-source-id: 93408951de0bb26d29fbc6170d208259f5148764