mark breakpoints as verified when they are set
Summary: HHVM stopped loading units in order to resolve breakpoints when they are set because it may not be safe to do so. Moreover this only works when the client sends full path of the file. This change left the breakpoints in "unverified" state which caused a lot of confusion for the Hack developers. This diff sets the "verified" field to true when a breakpoint is set. This is hacky, but better than loading the unit. One side-effect of this change is that when the breakpoint actually gets resolved, it might change the location (for example for multi-line statements).

Reviewed By: ebluestein

Differential Revision: D31808196

fbshipit-source-id: 4e57c1b235f48d2341933e7ec725e2243363124d