HH\Map::zip() has write_props coeffects
Summary:
`HH\Map::zip()` calls `next()` on two iterators, cranking them and mutating their state. Pure code can't mutate state.

One of these mutations is kind of fake - we create an iterator by calling `$this->getIterator()`, we crank it a few times, and then we throw it away. Pure code should be able to do this! It should be fine to create, mutate, and destroy an object within a pure function as long as nobody else ever sees it. This is the mutation that's throwing a violation.

The other mutation isn't caught yet - `zip()` accepts a `Traversable`, which could be an iterator or a generator. In that case, `zip()` mutates the `Traversable`, cranking it and mutating its state, making `zip()` impure.

So this is a weird situation where the mutation that we notice is something we shouldn't have noticed, and where the mutation that we didn't notice is something we should have noticed. Adding `write_props`, to be honest about the mutations that `zip()` does, will solve both problems.

Reviewed By: oulgen

Differential Revision: D31710240

fbshipit-source-id: 56cfb4f16877238fb9bd7e9f06a2d530fec900c2