fix refinement of like types
Summary:
Previously, the logic for refining a downcast with a known type parameter was not working properly for refining types of the form `dynamic | t`. With sound dynamic no refinement was ever performed (not flexible enough), and with current `dynamic` refinement was always performed.

```
class C<T> {}
<<__SupportDynamicType>> class D<T> extends C<T>
function f(~C<int> $c) : ~D<int> {
  $d = $c as D<_>;
  // with SDT, this was refining to D<T#1> & ~C<int>, which is less precise than
  // ~D<int>, which is also sound.
  // without SDT, this was refining to `D<int>` which is unsound.
  return $d;
}
```

Here, we detect the case and distribute the intersection over the union to get a `(D<T#1> & dynamic) | D<int>` result in general.

Reviewed By: andrewjkennedy

Differential Revision: D31544440

fbshipit-source-id: a44d8b87e6e328d9d61ac86a81af2ed87081aa0f