Interpret pessimization log
Summary:
Adds a new entry point to the `Pessimise`  module named `targeted_pessimise`. Targets for pessimisation are read from the logs and stored as keys in a set. The key format for a parameter or property `$p` declared in `/file/f.php` at line *n* is `/file/f.php,n,$p`.

This means that it is sufficient to operate exclusively over `PropertyDeclaration` or `ParameterDeclaration` nodes and use their stored positions. This breaks down if two functions are declared on the same line and share a parameter name that is a pessimisation target, but I don't think that case is worth worrying about. On the other hand, this allows us to ignore namespacing and avoid tracking the enclosing classes for methods and properties.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D31787758

fbshipit-source-id: f3d62b7ac84ad49db38ece84676a011f14bd25f2