Fix perf bug guard in irgen-bespoke.cpp
Summary:
The purpose of this check is to ensure that we don't have a performance bug in the analysis of bespoke array types. The types from bespoke irgen should be at least as precise as the types from vanilla irgen, modulo layout information, which may differ between the two.

The bug here is that we don't drop the layout if the type is, say, Dict=Vanilla|Null, because it's not a subtype of TArrLike. We should drop the array specialization whenever we have it.

Reviewed By: ricklavoie

Differential Revision: D31866396

fbshipit-source-id: df1730dc81ec952088832fa5a92d4b122fb68895