Share parts of catch blocks within the same inline function
Summary:
When in a catch block of an inlined function, instead of spilling all pending
inlined frames using InlineCall instructions, free the current stack and jump
to the shared block responsible for executing the rest of the EndCatch logic.
Instead of spilling the frame, the shared block frees it and exits it using
the EndInlining instruction. If the caller is another inlined frame, the whole
logic is applied recursively, otherwise the regular EndCatch instruction is
used.

This should free significant portion of the cold and frozen sections.

Reviewed By: ricklavoie

Differential Revision: D31855228

fbshipit-source-id: 06bd71ec83363ddd5497d481b2bfd5a896b412a0