Pure DateTime::getTimestamp
Summary:
HHVM special cases DateTimeInterface for relational comparison and equality checks.

https://www.internalfb.com/code/fbsource/[8f028692d25aec30c7d7c9602faa3feb8f4fbd83]/fbcode/hphp/runtime/base/object-data.cpp?lines=835-838%2C922-925

This means that object comparisons in pure contexts can cause VM re-entry and invoke `getTimestamp` on the object. Removing this operator overload would be ideal, but that requires a migration. However, looking at the implementation for `DateTime::getTimestamp` we can see it is pure, and marking the methods as pure may be good enough for purity purposes.

HHVM impl for `DateTime::getTimestamp`:
https://www.internalfb.com/code/fbsource/[98bb72d0e7440417f232c3fd8cb96b390f87796f]/fbcode/hphp/runtime/ext/datetime/ext_datetime.cpp?lines=226

Reviewed By: viratyosin

Differential Revision: D31864666

fbshipit-source-id: 0f6492959a23351bd4852b2a5c16e79fdfda5818