Refactor typing_deps.Dep to use compound hashes
Summary:
We used to construct dep edges containing both a type and a member using the function

```
Hash(Type, Member)
```

Unfortunately this means if we have a Hash(Type) and a member, tragically, we can't construct the dep edge Hash(Type, Member). We then have to do a bunch of funny business converting dep edges back to class names in an effort to construct... more dep edges!

jakebailey came up with the genius idea of using compound hashes to fix this. The idea is, what if we instead constructed dep edges using

```
Hash(Hash(Type), Member)
```

Then if we have a Hash(Type) and a member, we can indeed construct the dep edge!

That's what this diff does. It also fixes all call sites to use the new API

Differential Revision: D31762703

fbshipit-source-id: 62a595c62f1248f5fb5173ba48a2dd571a0ba97a