Fix bug in interaction between write_props check and readonly
Summary: The purity write_props check in the typechecker checks for array appends like [$write_props] by making sure the type of the value is a dict, keyset or vec. Unfortunately it doesn't handle readonly expressions correctly, which results in a bug of readonly dicts/vecs not being able to be appended to in a pure function.

Reviewed By: oulgen

Differential Revision: D31916476

fbshipit-source-id: b633206e85c051bd4be32bfa0d90f5a77e190102