Do not update VM regs after spillInlinedFrames()
Summary:
This logic was copied from the previous store-elim optimization, but is not
needed except for EndCatch. Move that logic out of spillInlinedFrames().

After a recent frame state logic change, spillInlinedFrames() is capable of
tracking which frames are spilled, so there is no harm calling it again. Remove
a special InterpCatch mode, as it is no longer necessary.

Use fp(env) and curSrcKey(env) instead of fixupFP(env) and fixupSrcKey(env)
after spillInlineFrames(), as it makes them equivalent.

Reviewed By: oulgen

Differential Revision: D31907563

fbshipit-source-id: ea678de1cd33685c6e7f5ade5c8e4582dafb0d5e