Make format strings magic work with like types in more positions
Summary:
The printf "magic" functions expect to find a declared method of a specific format on the interface I in `HH\FormatString<I>` and then deconstruct that type to build a new function type that is used to typecheck the call. This isn't robust against some pessimizations. Because it's sensitive to the shape of the declaration, it's simpler to strip off these pessimizations before doing the analysis.

This diff just extends the de-pessimization helpers to be more robust against mixing ~ and ? and handles the case where the parameter of `format_s` is a `~string`

Reviewed By: andrewjkennedy

Differential Revision: D31862645

fbshipit-source-id: 5df2581b89a2f4b292f50e38dc6f4237a9459c30