Add methods to send payloads to server and process credits
Summary:
Added two methods that client will use to interact with sink bridge. Client will first wait for credits and/or final response. Once it receives the credits, it will pass this to PHP thread via asio event.

If final response is not received, php thread will generate payloads using async generator. In order to prevent switching repeatedly between cpp and php, we will locally buffer the payloads in php thread depending on the number of credits received. If the server has still not cancelled then we will send the payloads and then wait for credits again.

Reviewed By: iahs

Differential Revision: D30723468

fbshipit-source-id: db052625ff8a4efabd3f39b4c342e19d69ffa2ec