Interpret reserved words as names in direct decl parser
Summary:
It's an error to use a reserved word where a name is expected, so the direct decl parser often doesn't handle it well. For instance, if you write `function using() {}`, the direct decl parser will produce a function decl with name `"\\"`, since it ignores `using` tokens.

Instead of dropping all information about keywords that the direct decl parser doesn't need (on correct programs), track it, so that we can emit decls that better represent invalid programs (e.g., a function decl with name `"\\using"`).

Reviewed By: bobrenjc93

Differential Revision: D31848460

fbshipit-source-id: 2c6b1ee5af73d3bb44b4356599e3e3f98e281d6c