Disallow object property writes in lambdas and add tests
Summary:
See D31347154 for more context. This diff handles the specific case where a lambda doing an object property write has an unspecified context but its enclosing function is pure (and therefore it itself is pure).

This is accomplished by falling back on an inherited `is_pure` value previously set by the enclosing function by making use of the fact that `visit_fun_def` is only visited for functions, but `visit_fun` is visited for both lambdas and functions.

Diff also adds tests for the cases when the lambda explicitly specifies not pure contexts, explicitly specifies pure context, and inherits from both pure and not pure contexts.

Reviewed By: jamesjwu

Differential Revision: D31590663

fbshipit-source-id: 7465a650e763b80391593d196d9c286e8d948dad