Fix internal error for Generators during SDT check
Summary:
The type inference code for `yield` aims to construct a `Generator<t, u, v>` type which is then subtyped with the declared return type. The `v` argument corresponds to the `send` method of Generator. Previously, it would copy the type from the declared return type, and raise an internal error if the type wasn't Generator otherwise. Now, if the return type is dynamic from an SDT pass, we say that the `send` argument is dynamic.

Now, the return check becomes `Generator<t, u, dynamic> <D: dynamic` which works when we pessimise the HHI (note how it's safe to pass dynamic to `send`)

```
+<<__SupportDynamicType>>
 final class Generator<+Tk, +Tv, -Ts> implements KeyedIterator<Tk, Tv> {
   ...
-  public function send(?Ts $v): void {}
+  public function send(~?Ts $v): void {}
   ...
 }
```

**Note:** A hackier alternative here is to generate `mixed` instead of dynamic for the `Ts` argument during the SDT pass, which would silence the error without pessimising the HHI.

Reviewed By: andrewjkennedy

Differential Revision: D31901983

fbshipit-source-id: 0c3cdb44c6f524b6ded93f4f853198149667cb2f