explicitly set Lwt engine
Summary:
Lwt defaults to using libev if it's available. Historically, we have not compiled hack with it available so have fallen back on the `select`-based engine. To enable testing, we need to be able to compile Lwt with libev, so this diff explicitly uses the select engine pending confirmation that libev is ok.

In flow, we do this in `LwtInit.run_lwt`. While I was porting this to hack (as `Lwt_util.run_main`), I also copied our fix for stack traces (see D19265118)

Reviewed By: shayne-fletcher

Differential Revision: D31928984

fbshipit-source-id: 3de1893951959bbdcf59d541491dafebc0176647