Make bidirectional type inference work for collections when the expected type is a like type
Summary:
The title is a bit misleading: the problem really is that the helper functions
- get_value_collection_inst
- get_key_value_collection_inst
- get_vc_inst
- get_kvc_inst

that are called when typing a `dict[...]`/`vec[...]` literal will try to pattern match the `expected` type -- usually the return type of the current function. This is for performance reasons as it is cheaper to check if all elements of the collection are subtypes of a target type than it is to build a union of all the elements and then check subtyping. It also serves to pass down the value type as the `expected` for the elements, which gets used for bidirectional type inference.

In this case, the helpers weren't extracting a value type because a darray became ~darray, and so inference could not infer lambda parameters. I had left a TODO the last time I ran into this problem; removed the TODO and confirmed all of the callsites of `expand_expected_and_get_node` should work with like types. In fact, that function already strips off `?` so this isn't a big stretch.

Reviewed By: andrewjkennedy

Differential Revision: D31912289

fbshipit-source-id: a085ddd50915c9e4dda8c077bbef2de0ca562b85