Extend facts_parser_www_perf to test facts migration
Summary:
Break `facts_parser_www_perf/run.rs` into 2 parts via subcommands `profile` and `test`
- `profile` mode is the existing code that profiles the facts parser on www
- `test` is added in this diff that runs the facts parser, then the direct decl parser, and compares the facts produced from the two. Add an `--input-file-list` argument to this subcommand, so we could control the size of the tests. Eventually we will want to test the whole www repo, but it's too inefficient for now.

Note that the `test` subcommand is temporary for the purpose of testing the facts migration from facts parser to the direct decl parser. Once the migration is complete, this script will be changed to profiling facts using the direct decl parser, and the testing code will be removed.

Reviewed By: shayne-fletcher

Differential Revision: D31850351

fbshipit-source-id: c09833e3aba9b3046101ae2fba7e0fc9a22bccc4