SrcKey: add a FuncEntry position
Summary:
Functions currently start their execution either at an offset 0, or an offset
corresponding to a DV initializer. It is also possible to reach these offsets
by other means, such as DV initializer jumping to the offset 0, or jumping from
other locations (although EntryNop opcode is designed to prevent that).

Mark SrcKeys corresponding to function entries with a FuncEntry tag and add
a special logic to execute code at this position when entering a function.

Future diffs will move some of the prologue logic to func entry, make func
entries callable rather than jumpable, and make it possible to execute rest of
the prologue logic at the caller side if the callee is known.

Reviewed By: ricklavoie

Differential Revision: D28976677

fbshipit-source-id: 683778f7ad0a02002b83c586b02c8a30aa4b78b0