Store in shmem only the first declaration of a given symbol in a file
Summary: Previously, we'd write all decls to shmem in order. If a symbol was declared multiple times in the same file, this would have the effect of overwriting the first declaration with the last. We want the first declaration to be the canonical one, so filter out repeated declarations before storing decls in shmem.

Reviewed By: bobrenjc93

Differential Revision: D31773136

fbshipit-source-id: 43be48fc62c7642fdab01ecce43f0cb7dcc85c3e