Expand dynamic for inference of collections and lambdas
Summary:
When typing a vec-like literal and the expected type is `dynamic`, treat the expected type as `Traversable<dynamic>` (passing the dynamic down into the inference of the vec elements). For dict-like literals, interpret as `KeyedTraversable<arraykey, dynamic>`. For closures, type the body under the expected type `(function (dynamic, ..., dynamic): dynamic)`.

This allows the following code to typecheck
```
function f(): dynamic {
  return ($p) ==> { $p->methCall() };
}
```
where previously there would be an error on the method call because the type of $p could not be inferred.

Reviewed By: jamesjwu

Differential Revision: D31945546

fbshipit-source-id: 5122010b2054b6675e991366d576557569ccdb6f