Fix Block::remove_if
Summary:
Block::remove_if was just delegating to InstructionList's remove_if
function. However if we remove an IRIntruction at the end of the
block, any Edges it might possess won't be removed. This can cause us
to have duplicate preds later on.

Instead iterate over the list ourself, and use the correct erase
function, which unhooks everything properly.

Add a check for duplicate preds to catch this kind of thing right
away.

Reviewed By: paulbiss

Differential Revision: D32008065

fbshipit-source-id: 5d487dc23fbdeeaeff5ab62d270df1d684fde7fc