Fix readonly recursion
Summary:
When recursing on statements in visitors, Rust's visitor framework does not visit the node itself. Thus we have to call visit directly when the thing we're recursing on is an expression or single statement we care about.

This matters if an if statement or while loop has exactly a single function call in it, i.e. `if(C\is_empty($foo))`
For safety this diff is gated to only the typechecker first, and I'll have a followup diff to turn it on in HHVM once the typechecker is released

Reviewed By: oulgen

Differential Revision: D32042335

fbshipit-source-id: 06ce7304f6510cfc9f35ab609212eeb8b0c70445