Save a byte in StructDict's perfect hash table entries
Summary:
Today, we use a StructDict's perfect hash table to do gets and sets without dereferencing its layout pointer. When we add type bounds on StructDict values, we'll need to store these bounds in the PerfectHashEntry, or else we'll lose this good caching behavior for sets.

The type bound will be represented in a single byte - it's a simple mask that we can use to test a type to see if it's valid. We can make room for this mask in PerfectHashEntry by replacing the (2-byte) type offset with the (1-byte) slot.

Reviewed By: ricklavoie

Differential Revision: D32037757

fbshipit-source-id: 0ff8f1f09ca6ec55ad6382f8144f18091cee681f