Lazily set free header kinds and lazily compute start bits
Summary:
A lot of these operations are in the pretty hot path of malloc small size slow.
This moves them in the colder GC paths that are possibly never executed in a
request.  Also by making these changes it becomes safe to allocate a block of
memory, and return it in smaller subcomponents.

Reviewed By: ricklavoie

Differential Revision: D31837465

fbshipit-source-id: 6c4037a4e5b98983f6fc70a1adcd607cbb5d4dea