make sure to decompress naming sqlite as well
Summary:
Now that we've rolled out naming sqlite in hack/64, I can finally test end to end. I realize that we aren't decompressing the naming sqlite table and thus when we look for "hh_mini_saved_state_naming.sql" in the saved state, it's not there (hh_mini_saved_state_naming.sql.zsd" is though!).

This diff fixes that

Differential Revision: D32048874

fbshipit-source-id: 498f3b0c0d1058491ef8d219d4c4190cb29e7df8