Fix coeffects violations in collections.ns.php
Summary:
`HH\Iterable::getIterator()` is marked as pure, but not all the constructors it calls are marked as pure. This diff marks each constructor as pure, moving the constructors' impure operations into memoized methods named `impureInit()`.

Modifying a Collection, like a `Vector` or `Map`, requires the `write_props` capability. Most methods in this file have `[defaults]`, which includes `[write_props]`, but the methods which have `[ctx $fn]` are not marked as `[defaults]`.

Reviewed By: rodmk

Differential Revision: D32011598

fbshipit-source-id: 1c0e9120fa60f7d237bc9e1f0284206d49788d38