Fix x64 disassembly when running retranslate-all
Summary:
Run doing retranslate-all, we generate the code into temporary
buffers, then relocate it all afterwards. However, when we print the
x64 disassembly, we do so as the unit is generated. As a result, the
addresses all refer to the temporary buffers. This is mainly harmless
(unless you're trying to use the dump in conjunction with a core
dump). However it breaks symbol printing. When disassembling, pass
along an extra "adjustment" parameter, which can be used to restore
the IP to the correct value to allow for correct symbol printing.

Reviewed By: oulgen, paulbiss

Differential Revision: D32080137

fbshipit-source-id: 159e76f7c92ce7e2265d0bf6163cfce7515fd356