Make it possible to use functionEnterHelper from interpreted call
Summary:
This helper is currently invoked from the prologue, which is never used from
the interpreted call. However, the next diff will move the usage to the func
entry, which can be reached from interpreted FCall.

When a FCall is interpreted, it sets saved RIP to a retHelper, which is
responsible for loading return offset from the ActRec and spilling the return
value, which overwrites the ActRec.

However, intercept handler already performs this logic, so the ActRec no longer
cointains the return offset. Resume handler needs to be used instead.

Reviewed By: ricklavoie

Differential Revision: D32079829

fbshipit-source-id: 8471e49fcf0ee1191e9c8294bf523d4439099971