inline measurements in server log
Summary:
Currently we log to the hack server log a great many rows about cgroup file, anon, total. We do this right at the end of the event, summarizing the memory changes of each "stage" associated with the event.

This made it hard to tell for instance which particular part of hh_server work contributed to the memory increase: we knew the name of the CgroupProfiler stage, but it was hard to tell how that name relates to other messages in the server log.

This diff changes cgroup logging to be more similar to the existing heap size messages that pepper our logs, recording the cgroup at the exact time at which it's measured, so you can recognize where it is in relation to all the other events.

This is how our existing heap-size measurements get recorded:
```
[2021-10-21 07:55:20.962] [master] Heap size: 0.001027G
```

And after this diff, here's a selection of examples of how cgroup measurements get recorded:
```
[2021-10-21 10:54:58.347] Cgroup: 2.42GiB  [startup/01_initialize...]

[2021-10-21 10:54:58.348] Cgroup: 2.42GiB  [startup/01_initialize end]

[2021-10-21 10:48:01.129] Cgroup: 2.50GiB  [startup/02_run start...]

[2021-10-21 10:48:01.129] Cgroup: 2.80GiB  [startup/02_run end]  --  high water mark 3.75GiB
```

Note: I also switched cgroup logs away from measuring in K/M/G to always measuring in GiB. I think that (1) GiB powers of 2 are more useful that G powers of 10 since that's how we normally characterize memory usage, and (2) it's nicer to stick with always GiB rather than jumping from MiB to GiB since that's the right order of magnitude for what we as a hack team should be interested in.

Differential Revision: D31822486

fbshipit-source-id: 006067fa0109d91da916e0bb90a412cb49b5e8ef