print at end of profile_memory
Summary:
[THIS DIFF IS SUPERSEDED BY ONE FURTHER UP THE STACK]

Here's part of the signature of CgroupProfiler:
```
val profile_memory :
  event:[ `Init of string | `Recheck of string ] -> (Profiling.t -> 'a) -> 'a

val print_summary_memory_table : event:[ `Init | `Recheck of string ] -> unit
```
The behavior is:
1. You call `profile_memory ~event` around a callback. All stages encountered during the callback will be considered part of the event.
2. At any time during that callback or after, you can call `print_summary_memory_table` to print either a completed event, or one underway. It'd be quite weird to call one currently underway since doing so would remove it from the Results list. Anyway, if you pass `Init` then it will print all `Init` events that have been started so far and remove them from the list; in practice there's only ever one. If you pass `Recheck` then it will print the recheck event of that name.

All this is quite complicated. I've simplified it a lot by just asserting that logs are printed at the end of `profile_memory`, end of story. This lets us do away with `Init | Recheck`. It also lets us do away with a global mutable list of which events are currently underway.

AND FURTHER UP THE STACK I CHANGE IT FURTHER.

Further up the stack, I change it to no longer use an end-of-event summary. Instead, each time it takes a cgroup measurement, it prints a line to the server log.

Differential Revision: D31821207

fbshipit-source-id: 7a5d4a2a7ae91d87d7e85429b9f83c483f0e4281