Set up some CI with human readable output for initial debugging
Summary:
This diff sets up a continuous integration test job for decls in comp. In the previous diffs, we have set up a way to ask the `hh_single_compile_cpp` binary to feed a file's decls into that file's compilation.

In this diff, we simply add the functionality that, when the compilation process asks for a decl, we will output the request as well as whether or not the request was successful in getting a decl.

Then, we augment compile.rs to simply request a known decl in the test file. Then, our test binary simply compares the output decl requests with the expected decl requests.

Reviewed By: shayne-fletcher

Differential Revision: D31842297

fbshipit-source-id: 7ac546e4ee516050aae3aadca0e3dcf4d5853976