Provide single direct decl API in parser/api
Summary: An API for direct decl parsing is provided in parser/api, and wrapped by a slightly different API in the decl_rust target. Instead of providing an API in parser/api which is only used by the wrapper, move the wrapper logic into parser/api, making that the single API for direct decl parsing.

Reviewed By: bobrenjc93

Differential Revision: D32029807

fbshipit-source-id: f465db55611cae63f20e7ebfc0897ec0cacd6e4c