Fix hackfmt for context aliases
Summary: {D28979231 (https://github.com/facebook/hhvm/commit/ccb78305da1e1a030e148c91f386a1949a789405)} wrote the formatting logic for ContextAliasDeclaration to share that of AliasDeclaration, but the syntax differs in that `ctx_alias_as_constraint` is a SyntaxList and the `= [...]` is optional for context aliases. The `t env type_constraint` was throwing an exception that SyntaxList should not be handled directly (line 99 of hack_format.ml). This fixes that.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D32161597

fbshipit-source-id: e6d3b28464695fc6fc381d0498f75f4db39916ba