Add field `abstract_` to shallow decl class
Summary:
Currently the direct decl parser turns `trait` and `interface` into an abstract `class` if the modifier `abstract` is present. For the purpose of deriving facts from shallow decls, however, we want to retain the original class kind. As a workaround, this diff adds a field `abstract` to the shallow decl class definition to indicate a class's abstraction, but keep the original class kind.

Guide for reviewers: main logic change is in `direct_decl_smart_constructors.rs`

Reviewed By: shayne-fletcher

Differential Revision: D32160538

fbshipit-source-id: 99d3ef597770486b48b754e080ef21f4342a64db