Fix awaitedTypeFromSSATmp
Summary:
We were attempting to infer the awaitable's inner type from the return
type of the function it came from. This is only safe if the Call HHIR
op did not specify async eager return. If it did, it's ambiguous
whether the awaitable came from the eager path (and thus is already
unpacked), or from the slow path (where it would be correct). Limit
this inference to only when async eager return isn't specified.

Reviewed By: arnabde03

Differential Revision: D32171204

fbshipit-source-id: e376bd4e7b920aa4b34ad34e891abceac072ba28