Add tests for lambda inheriting context and specifying context within class method
Summary: Lambdas are functions in the ast, so if they are inheriting context from an enclosing method (*not* function), visit_method_() will set the enclosing method's purity and visit_fun_() will take care of inheriting the purity. This wasn't obvious to me when I re-read the coeffects check file so I made two small tests to further clarify.

Reviewed By: jamesjwu

Differential Revision: D31979526

fbshipit-source-id: 3a9341178e4f37143fde29d2d2edcf392f183e52