Profile and optimize accesses to globals
Summary:
If we're performing jumpstart, profile all accesses to globals. We'll
serialize metadata for the N most frequently accesses globals (by
default 200) in the jumpstart file. This metadata includes the global
name, whether it was usually present, and the union of all the types
it was (when present).

When consuming the jumpstart file, we'll allocate RDS slots for
globals stored within the file. When JITing code to access one of
those globals, we'll instead emit code to check the RDS slot first. If
the slot is initialized, we simply load the value. Otherwise,
depending on the type of access, we either go the failure case (as
normal), or use the slower normal access path. If profiling indicated
that the global is almost always present (>=98% of the time), we'll
side-exit if it's not. Finally, if we profiled a type for the global,
we can optionally emit a type check on it. If the type check fails,
we'll again side-exit.

Since globals can still be accesses via the global name-value table,
we need for it to stay in sync with the entries in RDS. For that
purpose, add the ability for an entry in the name-value table to point
at either a TypedValue, or a RDS Link. If the global is one of the
profiled ones, it will point at it's link.

Differential Revision: D32016795

fbshipit-source-id: 3e84bdba73c39b63b6b9269b6f6cc495e6735091