Minor style nits in irgen-types
Summary:
We would like to support verifying a (nearly arbitrary) type constraint on a (nearly arbitrary) value - in particular, we don't want to require values to be DataTypeSpecific.

As I was modifying this code, I wanted to fix some small style nits to make the code easier to read (IMO).

Reviewed By: mofarrell

Differential Revision: D32179499

fbshipit-source-id: 991bc7d3bfdce186414139b0ac0582f202b1d53c