Support verifyTypeImpl for non-DataTypeSpecific values
Summary:
We modify verifyTypeImpl to support input values that are not DataTypeSpecific. We do so in the simplest possible way: we break the input value's type into its component DataTypes, check which ones could possibly pass the type constraint, and emit type checks for each case.

This approach doesn't work for verifyPropType, because that helper captures some of the intermediate SSATmps computed during type-checks and assumes that they're valid after the full check. This assumption is now wrong, because some of those intermediates are specific to one branch or another of the full type-check. We fix this issue for properties by checking if the input there is DataTypeSpecific and emitting the generic fallback if not.

Reviewed By: mofarrell

Differential Revision: D32183693

fbshipit-source-id: b0806cdea47930fec5e603a9e8a9b1302c4adb2c