Fix frame state ctx merging
Summary:
If two frames with different ctx are merged, we did not clear ctx.

This situation was not possible before D29354257, because we always eagerly
loaded the context before any opportunity to execute control flow. After
D29354257, we may perform control flow due to guards before the context of
a closure is updated.

Since we may lose ctx, add tracking of context type, which may help us
determine the exact type when inlining.

Reviewed By: ricklavoie

Differential Revision: D32208751

fbshipit-source-id: ba967d00d7a55001588db3981e838d701ef1edff