Prefetch a blockable when returning from an async function
Summary:
colavitam and I investigated why asyncFuncRet is slow (~0.5% exclusive CPU). It turns out the vast majority of the cost is simply accessing the heap memory of the next WaitHandle - typically, an AsyncFuncWaitHandle of the next frame.

We can reduce this cost by optimistically emitting a prefetch instruction at the beginning of an async RetC.

Reviewed By: colavitam

Differential Revision: D32218903

fbshipit-source-id: 3551817fab87d3ee011d5e0e183432907a2a1315